+++ changes.md
@@ -1,7 +1,10 @@
 ---
 layout: page
-title: Changelog
+title: Change Log
+nav_order: 2
 ---
+
+# Change Log
 
 ## 0.5.3
 - Minor doc and code usability improvvements.